refactor: block standardization
