+++ src/bin.rs
@@ -18,6 +18,7 @@ mod macros;
 mod frontend;
 mod generate;
 mod integrate;
+mod indexer;
 mod poke;
 mod publish;
 mod runnner;
