+++ src/indexer/chains/mod.rs
@@ -1,0 +1,5 @@
+pub mod bitcoin;
+pub mod stacks;
+
+pub use bitcoin::standardize_bitcoin_block;
+pub use stacks::standardize_stacks_block;
