+++ src/types/data.rs
@@ -29,8 +29,8 @@ pub struct StacksBlockData {
 #[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct StacksBlockMetadata {
     pub bitcoin_anchor_block_identifier: BlockIdentifier,
-    pub bitcoin_genesis_block_identifier: BlockIdentifier,
     pub pox_cycle_index: u32,
+    pub pox_cycle_position: u32,
     pub pox_cycle_length: u32,
 }
 
@@ -47,7 +47,7 @@ pub struct BitcoinBlockData {
     /// The timestamp of the block in milliseconds since the Unix Epoch. The
     /// timestamp is stored in milliseconds because some blockchains produce
     /// blocks more often than once a second.
-    pub timestamp: i64,
+    pub timestamp: u32,
     pub transactions: Vec<BitcoinTransactionData>,
     pub metadata: BitcoinBlockMetadata,
 }
